Extract sidebar project and workspace terminal state
