+++ src/app/state/workspace-action-handlers.ts
@@ -1,12 +1,19 @@
-import { isLocalSessionPath } from '../../../shared/session-paths'
 import type { Project, View } from '../types'
 import type {
-  NonGitOpsView,
   UtilityView,
   UtilityViewReturnState,
   WorkspaceAction,
   WorkspaceState,
 } from './workspace'
+import {
+  getGitOpsReturnView,
+  getTerminalStateForNextView,
+  getTerminalVisibilityForSession,
+  migrateTakeoverOverride,
+  setTerminalVisibleState,
+  shouldMigrateTerminalVisibilityForOpenedThread,
+  shouldRestoreTerminalOnGitOpsClose,
+} from './workspace-terminal-state'
 
 export function isUtilityView(view: View): view is UtilityView {
   return view === 'settings' || view === 'extensions' || view === 'skills'
@@ -26,111 +33,6 @@ function createUtilityViewReturnState(state: WorkspaceState): UtilityViewReturnS
     takeoverVisible: state.takeoverVisible,
     gitOpsReturnView: state.gitOpsReturnView,
     selectedDiffFilePath: state.selectedDiffFilePath,
-  }
-}
-
-function migrateTakeoverOverride(
-  takeoverOverrides: Record<string, boolean>,
-  fromSessionPath: string | null,
-  toSessionPath: string,
-) {
-  if (!fromSessionPath || fromSessionPath === toSessionPath) {
-    return takeoverOverrides
-  }
-
-  if (!isLocalSessionPath(fromSessionPath) || isLocalSessionPath(toSessionPath)) {
-    return takeoverOverrides
-  }
-
-  if (!Object.hasOwn(takeoverOverrides, fromSessionPath)) {
-    return takeoverOverrides
-  }
-
-  const { [fromSessionPath]: override, ...remainingOverrides } = takeoverOverrides
-  return {
-    ...remainingOverrides,
-    [toSessionPath]: override ?? false,
-  }
-}
-
-function getGitOpsReturnView(activeView: View, fallback: NonGitOpsView): NonGitOpsView {
-  if (activeView === 'gitops') {
-    return fallback
-  }
-
-  return activeView
-}
-
-function getTerminalVisibilityForSession(
-  terminalVisibleBySession: Record<string, boolean>,
-  sessionPath: string | null,
-) {
-  return sessionPath ? (terminalVisibleBySession[sessionPath] ?? false) : false
-}
-
-function shouldRestoreTerminalOnGitOpsClose(state: WorkspaceState) {
-  return state.terminalVisible && (state.activeView === 'thread' || state.activeView === 'code')
-}
-
-function shouldMigrateTerminalVisibilityForOpenedThread(
-  state: WorkspaceState,
-  action: Extract<WorkspaceAction, { type: 'open-thread' }>,
-) {
-  if (
-    state.activeView === 'code' &&
-    state.selectedProjectId === action.projectId &&
-    (state.projectTerminalVisibleByProject[action.projectId] ?? false)
-  ) {
-    return true
-  }
-
-  if (
-    state.activeView !== 'thread' ||
-    !state.selectedSessionPath ||
-    state.selectedSessionPath === action.sessionPath
-  ) {
-    return false
-  }
-
-  if (state.selectedThreadId === action.threadId) {
-    return !Object.hasOwn(state.terminalVisibleBySession, action.sessionPath)
-  }
-
-  return false
-}
-
-function getTerminalStateForNextView(state: WorkspaceState, nextView: View) {
-  const getProjectTerminalVisible = () =>
-    state.selectedProjectId
-      ? (state.projectTerminalVisibleByProject[state.selectedProjectId] ?? false)
-      : false
-
-  if (state.activeView !== 'gitops') {
-    return {
-      terminalVisible:
-        nextView === 'thread'
-          ? getTerminalVisibilityForSession(
-              state.terminalVisibleBySession,
-              state.selectedSessionPath,
-            )
-          : nextView === 'code'
-            ? getProjectTerminalVisible()
-            : state.terminalVisible,
-      restoreTerminalVisibleOnGitOpsClose: state.restoreTerminalVisibleOnGitOpsClose,
-    }
-  }
-
-  if (nextView === 'gitops') {
-    return {
-      terminalVisible: false,
-      restoreTerminalVisibleOnGitOpsClose: state.restoreTerminalVisibleOnGitOpsClose,
-    }
-  }
-
-  return {
-    terminalVisible:
-      (nextView === 'thread' || nextView === 'code') && state.restoreTerminalVisibleOnGitOpsClose,
-    restoreTerminalVisibleOnGitOpsClose: false,
   }
 }
 
@@ -390,28 +292,6 @@ function closeGitOpsState(state: WorkspaceState): WorkspaceState {
         : null,
     selectedDiffFilePath: null,
     utilityViewReturnState: null,
-  }
-}
-
-function setTerminalVisibleState(state: WorkspaceState, visible: boolean): WorkspaceState {
-  if (!state.selectedSessionPath) {
-    if (!state.selectedProjectId) return { ...state, terminalVisible: visible }
-    return {
-      ...state,
-      terminalVisible: visible,
-      projectTerminalVisibleByProject: {
-        ...state.projectTerminalVisibleByProject,
-        [state.selectedProjectId]: visible,
-      },
-    }
-  }
-  return {
-    ...state,
-    terminalVisible: visible,
-    terminalVisibleBySession: {
-      ...state.terminalVisibleBySession,
-      [state.selectedSessionPath]: visible,
-    },
   }
 }
